separate random raycast tests
