+++ test/MeshBVH.test.js
@@ -3,7 +3,7 @@
 */
 
 import * as THREE from 'three';
-import { MeshBVH, acceleratedRaycast, computeBoundsTree, disposeBoundsTree, CENTER, SAH, AVERAGE, getBVHExtremes } from '../src/index.js';
+import { MeshBVH, acceleratedRaycast, computeBoundsTree, disposeBoundsTree, getBVHExtremes } from '../src/index.js';
 
 THREE.Mesh.prototype.raycast = acceleratedRaycast;
 THREE.BufferGeometry.prototype.computeBoundsTree = computeBoundsTree;
@@ -791,101 +791,5 @@ describe( 'Raycaster', () => {
 describe( 'BoundsTree API', () => {
 
 	it.skip( 'test bounds tree and node apis directly', () => {} );
-
-} );
-
-describe( 'Random intersections comparison', () => {
-
-	let scene = null;
-	let raycaster = null;
-	let ungroupedGeometry = null;
-	let ungroupedBvh = null;
-	let groupedGeometry = null;
-	let groupedBvh = null;
-
-	describe( 'CENTER split', () => runRandomTests( { strategy: CENTER } ) );
-	describe( 'Lazy CENTER split', () => runRandomTests( { strategy: CENTER, lazyGeneration: true } ) );
-
-	describe( 'AVERAGE split', () => runRandomTests( { strategy: AVERAGE } ) );
-	describe( 'Lazy AVERAGE split', () => runRandomTests( { strategy: AVERAGE, lazyGeneration: true } ) );
-
-	describe( 'SAH split', () => runRandomTests( { strategy: SAH } ) );
-	describe( 'Lazy SAH split', () => runRandomTests( { strategy: SAH, lazyGeneration: true } ) );
-
-	function runRandomTests( options ) {
-
-		beforeAll( () => {
-
-			ungroupedGeometry = new THREE.TorusBufferGeometry( 1, 1, 40, 10 );
-			groupedGeometry = new THREE.TorusBufferGeometry( 1, 1, 40, 10 );
-			const groupCount = 10;
-			const groupSize = groupedGeometry.index.array.length / groupCount;
-
-			for ( let g = 0; g < groupCount; g ++ ) {
-
-				const groupStart = g * groupSize;
-				groupedGeometry.addGroup( groupStart, groupSize, 0 );
-
-			}
-
-			groupedGeometry.computeBoundsTree( options );
-			ungroupedGeometry.computeBoundsTree( options );
-
-			ungroupedBvh = ungroupedGeometry.boundsTree;
-			groupedBvh = groupedGeometry.boundsTree;
-
-			scene = new THREE.Scene();
-			raycaster = new THREE.Raycaster();
-
-			for ( var i = 0; i < 10; i ++ ) {
-
-				let geo = i % 2 ? groupedGeometry : ungroupedGeometry;
-				let mesh = new THREE.Mesh( geo, new THREE.MeshBasicMaterial() );
-				mesh.rotation.x = Math.random() * 10;
-				mesh.rotation.y = Math.random() * 10;
-				mesh.rotation.z = Math.random() * 10;
-
-				mesh.position.x = Math.random() * 1;
-				mesh.position.y = Math.random() * 1;
-				mesh.position.z = Math.random() * 1;
-
-				scene.add( mesh );
-				mesh.updateMatrix( true );
-				mesh.updateMatrixWorld( true );
-
-			}
-
-		} );
-
-		for ( let i = 0; i < 100; i ++ ) {
-
-			it( 'cast ' + i, () => {
-
-				raycaster.firstHitOnly = false;
-				raycaster.ray.origin.set( Math.random() * 10, Math.random() * 10, Math.random() * 10 );
-				raycaster.ray.direction.copy( raycaster.ray.origin ).multiplyScalar( - 1 ).normalize();
-
-				ungroupedGeometry.boundsTree = ungroupedBvh;
-				groupedGeometry.boundsTree = groupedBvh;
-				const bvhHits = raycaster.intersectObject( scene, true );
-
-				raycaster.firstHitOnly = true;
-				const firstHit = raycaster.intersectObject( scene, true );
-
-				// run the og hits _after_ because in the lazy generation case
-				// the indices will be changing as the tree is generated and make
-				// the results will look different.
-				ungroupedGeometry.boundsTree = null;
-				groupedGeometry.boundsTree = null;
-				const ogHits = raycaster.intersectObject( scene, true );
-
-				expect( ogHits ).toEqual( bvhHits );
-				expect( firstHit[ 0 ] ).toEqual( ogHits[ 0 ] );
-
-			} );
-
-		}
-
-	}
 
 } );
